pyproject.toml
[tool.poetry]
name = "scrapy-folder-tree"
version = "0.1.3"
description = "A scrapy pipeline which stores files using folder trees."
authors = ["Panagiotis Simakis <sp1thas@autistici.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/sp1thas/scrapy-folder-tree"
keywords = ["scrapy", "scrapy-files-structure", "scrapy-folder-tree", "folder-tree"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Pytest",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Natural Language :: English",
    "Natural Language :: Greek"
]
packages = [{ include = "scrapy_folder_tree", from = "src" }]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/sp1thas/scrapy-folder-tree/issues"

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2"

[tool.poetry.dev-dependencies]
pytest-cov = "^4.0.0"
Pillow = "^9.4.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.1"

[build-system]
requires = ["poetry-core>=1.0.0", "pre-commit>=3.0.4"]
build-backend = "poetry.core.masonry.api"
